+++ lib/onnxruntime/ffi.rb
@@ -136,7 +136,13 @@ module OnnxRuntime
         :ModelMetadataGetDescription, callback(%i[pointer pointer pointer], :pointer),
         :ModelMetadataLookupCustomMetadataMap, callback(%i[pointer pointer pointer pointer], :pointer),
         :ModelMetadataGetVersion, callback(%i[pointer pointer], :pointer),
-        :ReleaseModelMetadata, callback(%i[pointer], :void)
+        :ReleaseModelMetadata, callback(%i[pointer], :void),
+        :CreateEnvWithGlobalThreadPools, callback(%i[], :pointer),
+        :DisablePerSessionThreads, callback(%i[], :pointer),
+        :CreateThreadingOptions, callback(%i[], :pointer),
+        :ReleaseThreadingOptions, callback(%i[], :pointer),
+        :ModelMetadataGetCustomMetadataMapKeys, callback(%i[pointer pointer pointer pointer], :pointer),
+        :AddFreeDimensionOverrideByName, callback(%i[], :pointer)
     end
 
     class ApiBase < ::FFI::Struct