+++ test/onnxruntime_test.rb
@@ -232,7 +232,7 @@ class OnnxRuntimeTest < Minitest::Test
   def test_modelmeta
     sess = OnnxRuntime::InferenceSession.new("test/support/model.onnx")
     metadata = sess.modelmeta
-    # assert_equal ({}), metadata[:custom_metadata_map]
+    assert_equal ({"hello" => "world", "test" => "value"}), metadata[:custom_metadata_map]
     assert_equal "", metadata[:description]
     assert_equal "", metadata[:domain]
     assert_equal "test_sigmoid", metadata[:graph_name]