+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.3.1 (unreleased)
 
 - Updated ONNX Runtime to 1.3.0
+- Added `custom_metadata_map` to model metadata
 
 ## 0.3.0 (2020-03-11)